requirements.txt
matplotlib==3.8.2
pillow==10.2.0
